Start extracting models from actions and fix jest setup for async tests
