+++ src/app/data-donation/actions.ts
@@ -2,69 +2,13 @@
 
 import { db } from "@/db/drizzle";
 import { conversationParticipants, conversations, donations, messages, messagesAudio } from "@/db/schema";
+import type {NewConversation, NewMessage, NewMessageAudio} from '@models/persisted';
+
 
 function generateExternalDonorId(): string {
     const id = crypto.getRandomValues(new Uint8Array)?.toString().substring(0,6);
     return id
 }
-
-type Participant = string;
-
-type ExternalDonorId = string;
-
-enum DataSourceValue {
-    WhatsApp = "WhatsApp",
-    Facebook = "Facebook",
-    Instagram = "Instagram",
-}
-
-enum DonationStatus {
-    NotStarted = "notstarted",
-    Pending = "pending",
-    Complete = "complete",
-    Deleted = "deleted",
-}
-
-interface DataSource {
-    id?: number,
-    name: DataSourceValue
-}
-
-interface Message {
-    id?: string,
-    wordCount: number,
-    timestamp: number,
-    sender: string,
-}
-
-interface MessageAudio {
-    id?: string,
-    lengthSeconds: number,
-    timestamp: number,
-    sender: string,
-}
-
-interface Conversation {
-    id?: string,
-    isGroupConversation?: boolean,
-    conversationId?: string,
-    dataSource: string,
-    messages: Array<Message>,
-    messagesAudio: Array<MessageAudio>,
-    participants: Array<Participant>,
-}
-
-interface DataDonation {
-    id?: string,
-    donorId: string,
-    status?: DonationStatus,
-    externalDonorId?: ExternalDonorId,
-    conversations: Array<Conversation>,
-}
-
-type NewConversation = typeof conversations.$inferInsert;
-type NewMessage = typeof messages.$inferInsert;
-type NewMessageAudio = typeof messagesAudio.$inferInsert;
 
 export async function addDonation(donation: DataDonation) {
     const { donorId, externalDonorId, status} = donation;
