+++ tsconfig.json
@@ -23,8 +23,9 @@
     ],
     "paths": {
       "@locales": ["./locales"],
-      "@components/*": ["./src/components/*"],
-      "@": ["./src"],
+      "@components/*": ["src/components/*"],
+      "@models/*": ["src/models/*"],
+//      "@": ["./src"],
       "@/*": ["./src/*"]
     },
     "target": "ESNext"
